Authorization updates
