+++ nova/src/Authorization/Http/Authorizers/Destroy.php
@@ -1,16 +1,16 @@
 <?php
 
-namespace Nova\Authorization\Http\Authorizers\Roles;
+namespace Nova\Authorization\Http\Authorizers;
 
 use Silber\Bouncer\Database\Role;
 use Nova\Foundation\Http\Authorizers\BaseAuthorizer;
 
-class Update extends BaseAuthorizer
+class Destroy extends BaseAuthorizer
 {
     public function authorize()
     {
         $role = Role::find($this->route('role'))->first();
 
-        return $role && $this->user()->can('update', $role);
+        return $role && $this->user()->can('delete', $role);
     }
 }
