+++ nova/src/Authorization/Http/Controllers/RoleController.php
@@ -19,18 +19,18 @@ class RoleController extends Controller
         $this->middleware('auth');
     }
 
-    public function index(Authorizers\Roles\Index $auth)
+    public function index(Authorizers\Index $auth)
     {
-        return app(Responses\Roles\Index::class)
+        return app(Responses\Index::class)
             ->withRoles(Role::orderBy('title')->get());
     }
 
-    public function create(Authorizers\Roles\Create $auth)
+    public function create(Authorizers\Create $auth)
     {
-        return app(Responses\Roles\Create::class);
+        return app(Responses\Create::class);
     }
 
-    public function store(Authorizers\Roles\Store $auth, Requests\Roles\Store $request)
+    public function store(Authorizers\Store $auth, Requests\Store $request)
     {
         $role = dispatch_now(new Jobs\CreateRoleJob($request->validated()));
 
@@ -39,13 +39,13 @@ class RoleController extends Controller
         return $role->fresh();
     }
 
-    public function edit(Authorizers\Roles\Edit $auth, Role $role)
+    public function edit(Authorizers\Edit $auth, Role $role)
     {
-        return app(Responses\Roles\Edit::class)
+        return app(Responses\Edit::class)
             ->withRole($role);
     }
 
-    public function update(Authorizers\Roles\Update $auth, Requests\Roles\Update $request, Role $role)
+    public function update(Authorizers\Update $auth, Requests\Update $request, Role $role)
     {
         $role = dispatch_now(new Jobs\UpdateRoleJob($role, $request->validated()));
 
@@ -54,7 +54,7 @@ class RoleController extends Controller
         return $role->fresh();
     }
 
-    public function destroy(Authorizers\Roles\Destroy $auth, Role $role)
+    public function destroy(Authorizers\Destroy $auth, Role $role)
     {
         $role = dispatch_now(new Jobs\DeleteRoleJob($role));
